Tutorial3: simplify structure
- rename runTestDetector[name] -> [name]
- rename .tpl files to .h and include them where they are used
